rating: use signed parameters for rb_rating_render_stars (bug #631698)
Sometimes we need to render to negative coordinates (when scrolling a
tree view up, for instance), so they need to be signed.

+++ widgets/rb-rating-helper.c
@@ -161,10 +161,10 @@ gboolean
 rb_rating_render_stars (GtkWidget *widget,
 			GdkWindow *window,
 			RBRatingPixbufs *pixbufs,
-			gulong x,
-			gulong y,
-			gulong x_offset,
-			gulong y_offset,
+			int x,
+			int y,
+			int x_offset,
+			int y_offset,
 			gdouble rating,
 			gboolean selected)
 {
